requirements.txt
# tested on python=3.8.13
pandas==1.4.3
numpy==1.23.1
scikit-learn==1.1.1
psycopg2-binary==2.9.3
scipy==1.8.1
fastapi==0.79.0
uvicorn==0.18.2